+++ stacActions.config.js
@@ -1,0 +1,5 @@
+// import StacMap from './src/actions/stac/StacMap.js';
+
+export default {
+  // StacMap,
+};